Laying a Lawyer's Versioning breadcrumb
The original MIT Tech Licensing Office Mutual Agreement form.
http://web.mit.edu/tlo/www/downloads/pdf/NDA_Mutual.pdf

The thought is that MutualForm.md is the “current” version.  Of course
github has all the past versions, but there may be reason for this
other naming convention, too.  Note that software distributions also
tend to have a this kind of numbered versioning.